Add output of the result set (if any) when executing queries in the RUN command.
The PostgreSQL mechanism for displaying NOTICE or WARNING
messages is insufficiently implemented in psycopg2. The
same seems to hold true for other drivers, following the
Python DB-API 2.0 specs. Only the old style PyGreSQL
interface seems to provide proper access to it, but it
seems a bad choice to convert to that. So we have to live
without the NOTICE messages for the moment.